requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
python-dotenv==1.0.1
magic-filter>=1.0.9